Legacy migrations should check for null `migration` table and skip if a new installation.

+++ app/modules/power/migrations/0213_00_00_000001_add_columns_for_pmset_data.php
@@ -18,7 +18,7 @@ class AddColumnsForPmsetData extends Migration {
         $legacyVersion = $this->getLegacyModelSchemaVersion('power');
         $capsule = new Capsule();
 
-        if ($legacyVersion < static::$legacySchemaVersion) {
+        if ($legacyVersion !== null && $legacyVersion < static::$legacySchemaVersion) {
             $capsule::schema()->table('power', function (Blueprint $table) {
                 $table->string('hibernatefile')->nullable();
                 $table->string('schedule')->nullable();
@@ -102,7 +102,7 @@ class AddColumnsForPmsetData extends Migration {
     public function down() {
         $legacyVersion = $this->getLegacyModelSchemaVersion('power');
 
-        if ($legacyVersion == static::$legacySchemaVersion) {
+        if ($legacyVersion !== null && $legacyVersion == static::$legacySchemaVersion) {
             $capsule::schema()->table('power', function (Blueprint $table) {
                 $table->dropColumn([
                     'hibernatefile',
